feat: Implements sliceString interface on VigenereBreaker class
This commit implements the sliceString interface, which returns
a string from an original message, which each character in the
string begins at variable whichSlice and gets the next character
totalSlices away. This way you can return all letters which
were encrypted using the same key in the Vigenere cipher.

Additionally, this commit adds the VigenereBreakerTest class
which contains the sliceString test, which extensively tests
the sliceString method.