feat: Adds breakVigenere function
This commit adds the breakVigenere funtion to the VigenereBreaker
class. This function takes an encrypted message, the length of
the key, and the most common letter in the language, and uses
that to decrypt and return the original message.

Additionally, this commit adds the breakVigenere test to the
VigenereBreakerTest class, which tests this function with the
athens.txt and athens_keyflute.txt files.