feat: Adds readDictionary method to VigenereBreaker class
This commit adds the readDictionary method to the VigenereBreaker
class, which takes a FileResource object and iterates over all
the words, adds them to a hash map, and returns the hashmap.

Additionally, this commit includes a test for the readDictionary
method, which reads in English words from a file, and asserts that
the generated dictionary is the correct size and contains certain
English words.