feat: Implements countWords method in VigenereBreaker class
This commit implements the countWords method in the VigenereBreaker
class. This method takes a message and a HashSet representing the
language of the message's viable words. This method returns the
number of words in the message that are words in the language.

Additionally, this commit adds the countWords JUnit test in the
VigenereBreakerTest class.