feat: breakVigenere now void method which logs message to sys out
This commit changes the breakVigenere method in the VigenereBreaker
class to no longer take inputs, and log a decrypted message to
stdout. The user is prompted to supply a file containing the
encrypted message, and another file containing the languages
dictionary. The method now calls breakForLanguage in order to
decrypt the message. The decrypted message is then logged to
stdout.