feat: Adds mostCommonCharIn method to VigenereBreaker class
This commit adds the mostCommonCharIn method to the
VigenereBreaker class. This method takes a dictionary
representing all the words in a given language, iterates over
that dictionary, and return the most common character in that
language. Additionally, this commit includes a JUnit test for
this method, which tests it with both an English and
Portuguese dictionary.