feat: Adds breakForAllLangs method to VigenereBreaker class
This commit adds the breakForAllLangs method to the
VigenereBreaker class. This method takes and encrypted string
in an unknown language, as well as a HashMap of languages and
their word dictionaries, and returns an array containing the
detected language and decrypted message.

Additionally, this commit includes a test for the method in
the VigenereBreakerTest class. This test generates the languages
dictionary from all the languages in the
test/breaking_vigenere_cypher/dictionaries directory, and passes
the encrypted athens_keyflute.txt contexts with the language
dictionary to breakForAllLangs. Finally it asserts that the
decrypted message is correct and detected as English.